rename <ChannelPrunedLearner> to <ChannelPrunedLegacyLearner>
