+++ learners/channel_pruning_legacy/model_wrapper.py
