+++ learners/learner_utils.py
@@ -20,7 +20,7 @@ import tensorflow as tf
 
 from learners.full_precision.learner import FullPrecLearner
 from learners.weight_sparsification.learner import WeightSparseLearner
-from learners.channel_pruning.learner import ChannelPrunedLearner
+from learners.channel_pruning_legacy.learner import ChannelPrunedLegacyLearner
 from learners.channel_pruning_gpu.learner import ChannelPrunedGpuLearner
 from learners.discr_channel_pruning.learner import DisChnPrunedLearner
 from learners.uniform_quantization.learner import UniformQuantLearner
@@ -45,8 +45,8 @@ def create_learner(sm_writer, model_helper):
     learner = FullPrecLearner(sm_writer, model_helper)
   elif FLAGS.learner == 'weight-sparse':
     learner = WeightSparseLearner(sm_writer, model_helper)
-  elif FLAGS.learner == 'channel':
-    learner = ChannelPrunedLearner(sm_writer, model_helper)
+  elif FLAGS.learner == 'chn-pruned-lgc':
+    learner = ChannelPrunedLegacyLearner(sm_writer, model_helper)
   elif FLAGS.learner == 'chn-pruned-gpu':
     learner = ChannelPrunedGpuLearner(sm_writer, model_helper)
   elif FLAGS.learner == 'dis-chn-pruned':
